Switch out lambda-api for middy routing.
Middy routing is abstracted by an http-router internal library. A test-utils library is also added to help create the correct lambda events for the HTTP API proxy integration.
Explicitly state the stagename in the CFN template and import it as a environment variable in order to pass it to the routes when necessary.